+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 ## 2024-09-26
 
+### Added
+
+- You can now slot prisms, fragments, and fusions on your builds!
+
 ### Fixed
 
 - Fixed bug preventing some build edits from saving.